Smooth direction graph.
I remember an email I sent to the author.  I noticed that the direction graph
would often jump by 2pi when hovering around \pm pi.  Paulson's solution to this
was to add 2pi (or subtract 2pi, depending on the direction of the jump) until
the graph was more-or-less smooth again.