dirty commit
About to drive so committing to put this in the "cloud".

I started filling-in the main Paleo function.  This, of course, had some
repercussions to the other files.  I made _cpy functions for the shapes and
results.  I updated the Paleo test file to make sure some of my assumptions
hold.  As said, though, I haven't quite finished.